refs #2 Ref Proxy with new inject
The reference proxy protection now uses the new code injection. Also
added some unit tests to verify that the reference proxy protection is
not breaking anything.

The injection code got some additional features as this was the first
time it was required to inject an entire type and there were still a
few bugs in there.