Fixed constants protection
The constants protection was broken after fixing the marker service.
The change caused the injected code for decoding the constants to
become a legal target for the encoding phase of the constant
protection. This caused the number encoding to encode the values in the
inital decode function that is executed before the encoded data was
decompressed.