Solved reference proxy error
The strong mode of the reference proxy encountered an error, that caused the normal and the expression mode to fail at runtime. The reason for this issue was that the key decode multiplication was added randomly before or after the instructions that were used to generate the placeholder parameters. The traceservice correctly identified the "add" instruction as source for the parameter of the method, how ever adding the konstant directly before the add instruction modified the stack and caused a wrong addition to take place.

The issue is fixed, by extending the tracing to resolve all the instructions that take part in constructing the value that parameter. This way the new constant may be randomly placed before or after the entire stack that produces the parameter. This resolves the issue. The x86 was uneffected, because it always creates the call after the instructions that create the parameter and never before.

A unit test was added to verify the correct function of the reference proxy protection.